+++ src/Bridge/Native/MultitronFactory.php
@@ -18,6 +18,7 @@ use Multitron\Execution\ProcessExecutionFactory;
 use Multitron\Orchestrator\Output\ProgressOutputFactory;
 use Multitron\Orchestrator\Output\TableOutputFactory;
 use Multitron\Orchestrator\TaskOrchestrator;
+use Multitron\Tree\TaskNodeFactory;
 use Multitron\Tree\TaskTreeBuilderFactory;
 use Psr\Container\ContainerInterface;
 use StreamIpc\NativeIpcPeer;
@@ -48,6 +49,8 @@ final class MultitronFactory
     private ?TaskCommandDeps $taskCommandDeps = null;
 
     private ?TaskTreeBuilderFactory $taskTreeBuilderFactory = null;
+
+    private ?TaskNodeFactory $taskNodeFactory = null;
 
     public function __construct(private readonly ?ContainerInterface $container)
     {
@@ -98,6 +101,17 @@ final class MultitronFactory
     public function setTaskTreeBuilderFactory(?TaskTreeBuilderFactory $taskTreeBuilderFactory): self
     {
         $this->taskTreeBuilderFactory = $taskTreeBuilderFactory;
+        return $this;
+    }
+
+    public function getTaskNodeFactory(): ?TaskNodeFactory
+    {
+        return $this->taskNodeFactory;
+    }
+
+    public function setTaskNodeFactory(?TaskNodeFactory $taskNodeFactory): self
+    {
+        $this->taskNodeFactory = $taskNodeFactory;
         return $this;
     }