+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - optional $tags argument in TaskTreeBuilder
 - immutable setters in TaskNode
+- optional TaskNodeFactory
 
 ## [1.0.0] - 2025-11-07
 ### Added